requirements.txt
selenium-4.4.3
pytest-7.1.3